pyproject.toml
[tool.poetry]
name = "argononetwo"
version = "0.1.0"
description = ""
authors = ["Peter (Rocky) Villano <peter.s.villano@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
black = "^22.3.0"
fake-rpi = "^0.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
